+++ src/QtImageReader.cpp
@@ -102,7 +102,7 @@ void QtImageReader::Close()
 }
 
 // Get an openshot::Frame object for a specific frame number of this reader.
-tr1::shared_ptr<Frame> QtImageReader::GetFrame(int requested_frame) throw(ReaderClosed)
+tr1::shared_ptr<Frame> QtImageReader::GetFrame(long int requested_frame) throw(ReaderClosed)
 {
 	// Check for open reader (or throw exception)
 	if (!is_open)
