+++ src/effects/Brightness.cpp
@@ -75,7 +75,7 @@ int Brightness::constrain(int color_value)
 
 // This method is required for all derived classes of EffectBase, and returns a
 // modified openshot::Frame object
-tr1::shared_ptr<Frame> Brightness::GetFrame(tr1::shared_ptr<Frame> frame, int frame_number)
+tr1::shared_ptr<Frame> Brightness::GetFrame(tr1::shared_ptr<Frame> frame, long int frame_number)
 {
 	// Get the frame's image
 	tr1::shared_ptr<QImage> frame_image = frame->GetImage();
@@ -174,7 +174,7 @@ void Brightness::SetJsonValue(Json::Value root) {
 }
 
 // Get all properties for a specific frame
-string Brightness::PropertiesJSON(int requested_frame) {
+string Brightness::PropertiesJSON(long int requested_frame) {
 
 	// Requested Point
 	Point requested_point(requested_frame, requested_frame);
