+++ src/effects/Deinterlace.cpp
@@ -59,7 +59,7 @@ void Deinterlace::init_effect_details()
 
 // This method is required for all derived classes of EffectBase, and returns a
 // modified openshot::Frame object
-tr1::shared_ptr<Frame> Deinterlace::GetFrame(tr1::shared_ptr<Frame> frame, int frame_number)
+tr1::shared_ptr<Frame> Deinterlace::GetFrame(tr1::shared_ptr<Frame> frame, long int frame_number)
 {
 	// Get original size of frame's image
 	int original_width = frame->GetImage()->width();
@@ -146,7 +146,7 @@ void Deinterlace::SetJsonValue(Json::Value root) {
 }
 
 // Get all properties for a specific frame
-string Deinterlace::PropertiesJSON(int requested_frame) {
+string Deinterlace::PropertiesJSON(long int requested_frame) {
 
 	// Requested Point
 	Point requested_point(requested_frame, requested_frame);
