+++ tests/ReaderBase_Tests.cpp
@@ -41,7 +41,7 @@ TEST(ReaderBase_Derived_Class)
 	public:
 		TestReader() { };
 		Cache* GetCache() { return NULL; };
-		tr1::shared_ptr<Frame> GetFrame(int number) { tr1::shared_ptr<Frame> f(new Frame()); return f; }
+		tr1::shared_ptr<Frame> GetFrame(long int number) { tr1::shared_ptr<Frame> f(new Frame()); return f; }
 		void Close() { };
 		void Open() { };
 		string Json() { };
